refactor: rename editors file to agent_editors
